# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact: {}
  description: API for managing video streams, uploads, and analytics powered by Mux
  title: NNP Stream Backend API
  version: "1.0"
paths:
...
  /analytics/assets:
    get:
      description: Returns a paginated list of all Mux video assets
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List video assets
      tags:
      - analytics
  /analytics/assets/{assetID}:
    get:
      description: Returns details for a specific Mux video asset including duration,
        status, resolution, and playback IDs
      parameters:
      - description: Mux Asset ID
        in: path
        name: assetID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get asset details
      tags:
      - analytics
...
  /analytics/views:
    get:
      description: Returns a paginated list of video views within a timeframe
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: 24:hours
        description: Timeframe filter (e.g. 24:hours, 7:days)
        in: query
        name: timeframe
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List video views
      tags:
      - analytics
  /analytics/views/{videoViewID}:
    get:
      description: Returns detailed information for a specific video view session
      parameters:
      - description: Video View ID
        in: path
        name: videoViewID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get video view details
      tags:
      - analytics
...
swagger: "2.0"